Move helper functions to weblibrary
