+++ components/tools/OmeroWeb/test/integration/weblibrary.py
@@ -25,6 +25,7 @@ import library as lib
 
 from django.test import Client
 from django.core.urlresolvers import reverse
+from urllib import urlencode
 
 
 class IWebTest(lib.ITest):
@@ -71,3 +72,30 @@ class IWebTest(lib.ITest):
         assert response.status_code == 302
         cls.django_clients.append(django_client)
         return django_client
+
+
+# Helpers
+def _post_reponse(django_client, request_url, data, status_code=403):
+    response = django_client.post(request_url, data=data)
+    assert response.status_code == status_code
+    return response
+
+
+def _csrf_post_reponse(django_client, request_url, data, status_code=200):
+    csrf_token = django_client.cookies['csrftoken'].value
+    data['csrfmiddlewaretoken'] = csrf_token
+    return _post_reponse(django_client, request_url, data, status_code)
+
+
+def _get_reponse(django_client, request_url, query_string, status_code=405):
+    query_string = urlencode(query_string.items())
+    response = django_client.get('%s?%s' % (request_url, query_string))
+    assert response.status_code == status_code
+    return response
+
+
+def _csrf_get_reponse(django_client, request_url, query_string,
+                      status_code=200):
+    csrf_token = django_client.cookies['csrftoken'].value
+    query_string['csrfmiddlewaretoken'] = csrf_token
+    return _get_reponse(django_client, request_url, query_string, status_code)
